requirements.txt
# Broker & market data
alpaca-py>=0.21.0
yfinance>=0.2.40

# Technical analysis
pandas-ta>=0.3.14b
pandas>=2.1.0
numpy>=1.26.0

# NLP / sentiment
transformers>=4.40.0
torch>=2.2.0
sentencepiece>=0.2.0

# News feeds
finnhub-python>=2.4.19
feedparser>=6.0.11
websockets>=12.0

# Scheduling & time
apscheduler>=3.10.4
pandas-market-calendars>=4.4.0
pytz>=2024.1

# Config & environment
python-dotenv>=1.0.1
pyyaml>=6.0.1

# Utilities
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.9.0

# Optional: Telegram alerts
python-telegram-bot>=21.0

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
